requirements.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
asgiref==3.6.0
automium==0.2.6
automium-web==0.1.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-request==1.6.2
djangorestframework==3.14.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.2
parts==1.5.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.2
